pyproject.toml
[project]
name = "lightyear-portfolio-ai-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "groq>=1.0.0",
    "jinja2>=3.1.6",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.2",
    "supabase>=2.28.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:supabase",
]
